requirements.txt
openai
python-dotenv
streamlit

# library for images
pillow